+++ packages/expo-router/src/fork/getPathFromState.ts
@@ -1,6 +1,8 @@
 import { PathConfig, PathConfigMap, validatePathConfig } from '@react-navigation/core';
 import type { NavigationState, PartialState, Route } from '@react-navigation/routers';
 import * as queryString from 'query-string';
+import Constants from 'expo-constants';
+import { Platform } from 'react-native';
 
 import { matchDeepDynamicRouteName, matchDynamicName, matchGroupName } from '../matchers';
 
@@ -400,7 +402,7 @@ function getPathFromResolvedState(
     }
   }
 
-  return { path: basicSanitizePath(path), params: decodeParams(allParams) };
+  return { path: appendBasePath(basicSanitizePath(path)), params: decodeParams(allParams) };
 }
 
 function decodeParams(params: Record<string, string>) {
@@ -607,3 +609,16 @@ const createNormalizedConfigs = (
   Object.fromEntries(
     Object.entries(options).map(([name, c]) => [name, createConfigItem(c, pattern)])
   );
+
+function appendBasePath(
+  path: string,
+  // @ts-expect-error: pending https://github.com/expo/universe/pull/13294
+  assetPrefix: string = Constants.expoConfig?.experiments?.basePath
+) {
+  if (process.env.NODE_ENV !== 'development') {
+    if (assetPrefix) {
+      return `/${assetPrefix.replace(/^\/+/, '').replace(/\/$/, '')}${path}`;
+    }
+  }
+  return path;
+}