+++ packages/expo-router/src/fork/getStateFromPath.ts
@@ -2,6 +2,7 @@ import { PathConfigMap } from '@react-navigation/core';
 import type { InitialState, NavigationState, PartialState } from '@react-navigation/routers';
 import escape from 'escape-string-regexp';
 import * as queryString from 'query-string';
+import Constants from 'expo-constants';
 import URL from 'url-parse';
 
 import { findFocusedRoute } from './findFocusedRoute';
@@ -44,16 +45,22 @@ type ParsedRoute = {
   params?: Record<string, any> | undefined;
 };
 
-export function getUrlWithReactNavigationConcessions(path: string) {
+export function getUrlWithReactNavigationConcessions(
+  path: string,
+  // @ts-expect-error: pending https://github.com/expo/universe/pull/13294
+  basePath: string | undefined = Constants.expoConfig?.experiments?.basePath
+) {
   const parsed = new URL(path, 'https://acme.com');
   const pathname = parsed.pathname;
 
   // Make sure there is a trailing slash
   return {
     // The slashes are at the end, not the beginning
-    nonstandardPathname: pathname.replace(/^\/+/g, '').replace(/\/+$/g, '') + '/',
+    nonstandardPathname:
+      stripBasePath(pathname, basePath).replace(/^\/+/g, '').replace(/\/+$/g, '') + '/',
+
     // React Navigation doesn't support hashes, so here
-    inputPathnameWithoutHash: path.replace(/#.*$/, ''),
+    inputPathnameWithoutHash: stripBasePath(path, basePath).replace(/#.*$/, ''),
   };
 }
 
@@ -734,3 +741,12 @@ const parseQueryParams = (path: string, parseConfig?: Record<string, (value: str
 
   return Object.keys(params).length ? params : undefined;
 };
+
+function stripBasePath(path: string, assetPrefix: string) {
+  if (process.env.NODE_ENV !== 'development') {
+    if (assetPrefix) {
+      return path.replace(/^\/+/g, '').replace(new RegExp(`^${escape(assetPrefix)}/`, 'g'), '');
+    }
+  }
+  return path;
+}