+++ packages/expo-router/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### 🎉 New features
 
+- Add support for `experiments.basePath` and hosting from sub-paths.
 - Add static font extraction support with `expo-font`. ([#24027](https://github.com/expo/expo/pull/24027) by [@EvanBacon](https://github.com/EvanBacon))
 
 ### 🐛 Bug fixes